Remove facility when ICAT url can't be determined
During startup, if we request the ICAT url from the IDS for a facility
but the request fails, remove the facility from the configuration, and
attempt to continue (so long as there are more facilities).  This may
still not work, but now a user can continue to use facility A even if
facility B stops responding, which was not possible beforehand.