Change JavaEE code to expect facilityName instead of icat/idsUrl
REST calls (and Upload proxy) rewritten to expect a facilityName
parameter rather than icat/idsUrl. Each method now expects the supplied
facilityName to be mapped to an icat/idsUrl via a property called
facility.<facilityName>.icatUrl (or idsUrl).

The Javascript is as yet unmodified, so the system will not work at this
point.