Log several otherwise-ignored exceptions
There are at least four locations in the Java code where exceptions are
caught but then ignored, and usage elsewhere appears to rely on this.
Added logging for three cases.

The fourth known case is in Cache.setValue(), where any exception raised
while trying to serialize the supplied object is ignored and nothing
happens; in this case "all" that happens is that the value is not
cached.