Log and try to display 403 failures in browse view
A so-far failed attempt to improve behaviour in cases such as the
following:

- log into facility A, and go to the Browse tab
- log out of A, then log into facility B

The browse tab will stick on "Loading - please wait", with the items
count spinner icon.
The browser console should show that Topcat is sending queries to
facility A (presumably to refresh the Browse page), but these are
failing with a 403 as the user is no longer logged in.

Added error callbacks that log the response error.
Added code that attempts to change the browse message to "No permission"
or similar in this case.
However, (a) the exact mechanism for detecting this case is not clear;
(b) the "Loading" message still appears "underneath" the "No permission"
message, even though the flags are being set appropriately in the same
block;
(c) the console now sometimes shows a "can't build query" error, for no
obvious reason.

These changes *might* be on the road to a usable solution; but not yet!