Improve login failure message
Previoiusly, when a login request to an ICAT instance returned a 404,
topcat responded with an error message to the user of "Type error:
response is null".  This commit fixes the login case of issue #346;
other cases may remain and have still to be investigated.

In tc-icat.service.js, added a failure handler to .login's outermost
.post() promise; this logs the error to the console, and returns a
rejection object with a message field stating, "Unable to log into
<facility>; ICAT may be down".

In login.controller.js, changed .login()'s failure handler to cope when
it is supplied a null value.