Guard against null failure responses from REST requests
The fundamental REST failure handler assumed that it was being passed a
non-null response object. This may have been the reason for the
"TypeError: response is null" messages.
This change guards against that; but note that it will pass null to the
promise rejection, and it's possible that receivers will not be prepared
for this.