Ignore facility if icat properties request fails
During startup, if a request for a facility's ICAT's properties (to get
the list of authenticators) fails, remove the facility from the
configuration and attempt to continue. Previously, topcat would simply
have stopped completely. Now, it may be possible for users to continue
to work with other facilities. However, if a user has cart or download
items from the failed facility, they will probably hit further problems.

This is similar to the case where /ids/getIcatUrl fails.