Add libgconf2-dev to Vagrantfile
In local vagrant dev, protractor tests were failing with code 199.
Following suggestions in
<https://github.com/angular/protractor/issues/3760>, found that the test
browser was failing to start because it couldn't find libgconf2-dev.
Adding it fixed the problem. As the TravisCI build is failing at the
same point, perhaps adding the missing library to the Vagrantfile will
improve matters?