Javascript calculates transport URLs from topcat.json
Originally the transport URL was calculated from the transportUrl sent
in the download object from the Java side. This does not (or is not
guaranteed to) work when the Java side is configured with different
URLs.  Instead, we now use the download's facilityName and transport
(type) to look up the idsUrl in topcat.json.

This has been tested by making an obvious change to the http transport's
idsUrl in topcat.json, and confirming that this change appears in both
the iframe created for a cart download, and in the URLs in the Downloads
dialog's Download buttons.

This should be sufficient to fix the issue; but the Java side still sets
(and uses) the transportUrl field in the download object, and this is
still passed to the Javascript side, but now ignored. So the field
should probably be removed and replaced on the Java side by a similar
lookup using topcat.properties.