Disable upload feature in topcat.json.example
A previous commit fcd4684 ("disabled upload feature to simplify first time install") did this by removing the idsUpload* properties; but leaving enableUpload set to true; but in this case, Topcat still shows the Upload button, but it will not work. So instead, we retain/restore the idsUpload* properties, and set enableUpload to false.

This fixes #371; but documentation on how to configure for uploads needs to be written; and Topcat needs to be more rigorous in determining whether or not uploads are possible.