Info button changes
Fixes #375

Add the Info button to a new first column on its own.
Only add the Info button when browse metaTabs are defined for the entity.
To do this, added a showInfoButton parameter, set by callers.

Note: in the Search panel, we can't determine the facility at the time when the gridOptions are set up, and so cannot determine whether there will be metaTabs. In this case, we always show the Info button; if there is no metadata, clicking it will tell the user (but won't crash).