Pass row entity to action button's show method
This enables buttons to define a show() method that is entity-dependent.

The original motivation is from the ISIS plugin, where users should be allowed to change permissions on (only those) investigations for which they are the PI.

Note that the show() method is called many times, so the implementation should not be too intensive; direct ICAT service requests are probably a bad idea!

The original code called the show() method with no arguments. It appears that passing an argument does not affect existing implementations (e.g. in other plugins) that do not expect arguments.