Clear user cart cache on logout
The per-facility User object uses a cache to reduce the frequency of GET
requests to the UserResource for the cart; but there was no mechanism to
clear the cache when the user is logged out. Clearing the cache on
logout forces a GET request, which passes the sessionId. This prevents
the new user from seeing the old user's cart, and so fixes #384 (or at
least the most significant problem).