Pass null facilityName through in AdminResource
Some REST methods in AdminResource originally generated a ForbiddenException when passed an empty icatUrl; but the change to passing a facilityName led to an InternalException that masks this. Changed the code so that a null facilityName argument is converted to a null icatUrl for subsequent code to handle as before.
While doing this, spotted that the exception string in FacilityMap.getIdsUrl was incorrect.