DOI redirection: inform users when entity query fails
If the DOI entity query returns no results, or a null result, assume this is because the user cannot read the entity, and inform them. (Previously, this would result in a TypeError message.)
We could go to the default home page, but a "blank" page makes it more obvious that the request cannot be satisfied; the user could always click the Home button if they want to continue.

Fixes #388 .