Support translation of the inform message
Look up DOI.QUERY_EMPTY in the translations. If the value is found, replace any references to _ENTITY_TYPE_ with the entity type. If no value is found, use a hard-wired English default.

Added a QUERY_EMPTY string to lang.json; and the DOI translation properties to lang.json.example.