Make setup mail properties optional
If any of the mail.xxx properties in topcat-setup.properties are omitted, do not create the mail resource. (Previously, omission caused setup to fail.)
In this case, mail.enable MUST NOT be set to true in topcat.properties. This is not tested during setup. Added a note to the installation notes to point this out.
Fixes #389.