Allow topcat.json to set a lifetime for investigation size caches
If set, use the property investigationSizeCacheLifetimeSeconds to set an age limit on cache entries. The default value of 0 means "no age limit".

Though the cache service already supported specifying an age limit (in the get() method), this had to be modified as it had a serious shortcoming that meant that a get() request with 0 would *always* force recalculation of an entry that had previously had an age limit set.

During tests, it appeared that multiple threads are writing to the same cache asynchronously, so that multiple store requests are being made for the same key. This commit does not fix this, and by forcing more frequent updates may make things worse. Further performance testing is required.