Javascript: add filter function to cache; add config to not cache zero-sized investigations
Added an optional dontCache function to the Cache constructor. If defined, cache.get() will call it for a new key/value pair, and if it returns true, it will not store the value.
Modified tc-icat.service: if the new config property dontCacheZeroSizedInvestigations is true, then define an appropriate dontCache function and pass it to the Cache constructor.
Note that tc-icat.service uses the cache to store queries as well as getSize requests. One downside of this approach is that the dontCache function will be called for all query stores as well; but it should quickly evaluate to false so that queries are always cached. It does not seem to impede performance, though it has only been tested against LILS.