Add configurable monitoring of low-priority queue behaviour
Uncached getSize requests are put onto a low-priority queue. Added monitoring that reports max queue length and max wait time to the browser console at one-minute intervals. Figures are reported for "ever" (since monitoring was enabled) and for up to the last 10 minutes.
Added a setMonitoring(bool) method to the top-level topcat service that can be used to enable or disable monitoring from the browser console, e.g. "tc.setMonitoring(true)".
Monitoring is disabled initially, so it is not possible to collect stats for the initial page load (or explicit reload).