Add DELETE clearSize to AdminResource
AdminResource.clearCachedSize() uses a new method CacheRepository.remove().
I attempted to add a test to CacheRepository, but the test fails because the repository bean does not get set up. I don't know how to do that; and I note that all other tests that refer to the CacheRepository have all been commented-out.