Update more component versions
Updated the "obvious" component versions (those that matched angular 1.4.8). With this change, in Firefox's console, angular.version is reported as 1.5.11 (at present), and Topcat seems to work. I don't know which other components may be worthwhile updating.

Current thinking is that migration of the Topcat code from AngularJS to Angular is not worthwhile, as it is likely to require considerable redesign. So it is unlikely that much more will be done on upgrading. It may still be wise to move to AngularJS 1.7.