Replace template strings in tc-cache.service.js
Discovered that the jsdoc3 plugin cannot cope with template strings (used in console logs for cache monitoring) and fails to build the jsdoc pages. The plugin is no longer supported, so ideally we should replace it with something else; but it is simpler to replace the template string constructions with simple concatenations.