Added relationships between Shift and Instrument to icatSchema
I am not certain that this captures the new relationship between Shift and Instrument accurately; it is based in part on the pattern for Dataset/Datafile.
I believe that the variableName 'shiftInstrument' cannot be simplified to 'instrument' as that name is used in facility.relationships; but it may be OK to rename 'instrumentShift' to just  'shift' as that is not used anywhere else.