Add logging to the plugin load process
Added console logging to show what Topcat is doing when it adds plugins/scripts and waits for them to load. The process is quite convoluted!

If a plugin's startup misbehaves, this could cause Topcat to hang at the loading phase; but it is not clear how Topcat could guard against this. Perhaps the waitForPlugins and/or waitForPluginScripts methods could maintain iteration counts and at least log to console if these start to become unreasonably large.

This change should not be pulled into the master as it produces large amounts of logging in the browser console. But it may be useful to return to it if asked to investigate plugin load failures or hangs.