Allow authenticators to have their own login buttons
Added a showAsButton boolean property to authenticationTypes in topcat.json; when true, the authenticator is removed from the choice llist and replaced by its own button.

If any button-showing authenticator requires username/password, the input boxes are always shown (I trust the reason is obvious).

Known problem: the extra buttons are not disabled when the input boxes are empty. Clicking on a button that requires inputs results in an obscure/internal "could not satisfy overloaded method" error.

There are some subtle points that can arise when there is a mix of button and non-button authenticators:

  - if a button authenticator requires username/password, then any 'anon' or external authenticator **must** also have its own button. This should be a configuration error, reported during startup, but is not (yet).
  - if login via a button authenticator fails, the selected authenticator will be reset to one in the choice list (if this is not empty). Without this, a choice list of one authenticator could become unreachable.