Improve button-disabling logic
All buttons are now enabled for Anon/external authenticators, and otherwise disabled if either username or password is blank.
(Also removed a test 'return true' line in showCredInputs.)

In some cases, the input boxes can appear highlit in red even though they are not (always) required. I do not plan to attempt to fix this.