Fix login page case
When doi-redirection goes to the login page first, replace the redirection step in the browser history. Before this change, a user who decided not to log in could not get back to the original page from which the doi-redirection link was launched.

I have deliberately retained the extra console logging that revealed how the switch to the 'login' state was happening, as it may prove useful again in future. However, I will remove it in the next commit.

Fixes #424.