Replace the URL location on logout
The LogoutController always goes to the 'login' state. By setting location:'replace', it's now possible to get to earlier pages using the Back button.

This is not always intuitive: when the user goes back to earlier Topcat pages, as they are now logged out, Topcat replaces each of them with the login page; so it can take multiple Back button presses (which always show the login page) to get to pre-Topcat pages in the history. I doubt this can be improved unless we **never** keep Topcat browse history, which would not be a good idea!