Attempt: replace hardwired inclusions [Not Working]
See #405.

An attempt to replace the hard-wired inclusions by including the entityTypes discovered in the metatabs configuration.
This does not (always) work. In the current LILS configuration, metatabs for Investigation and Datafile are OK, but Dataset is not - the tabs do not complete properly, and sometimes there is a console error "Assertion failed: Input argument is not an HTMLInputElement".
I suspect that the hard-wired inclusions are still required. Whether also adding inclusions found in the configuration helps when "new" entity variables are mentioned is unknown.

Note that this change also adds logging for debug purposes.