Allow ids.timeout to be in milliseconds, seconds or minutes
Changed the parsing of ids.timeout: it can now have an optional "s" or "m" suffix. With no suffix, the value is interpreted as milliseconds - this is for compatibility with existing configurations. If the value is not set, or does not match the expected format, no (explicit) timeout is set.

FIxes #349.