Check status of download types, properly
The previously-committed code did not work: we need to take into account that the disabled-status checks are promises, which required a larger rewrite of download-cart's ok() button function. Note the large-ish comments added to the code.
In tc-user's getDownloadTypeStatus, added the forgotten facilityName parameter.
Added setDownloladStatusType to tc-admin. At the moment, this  was just used in a hard-wired call to disable the 'globus' download type (not committed). It could be used by a future Admin dialog to disable/enable download types.