Allow/cope with empty disabled message for download types
The API allows a download type to be disabled with an empty message. Changed download-cart.controller so that in this case a default message is disabled to the user.
Changed topcat_admin so that the user can keep the previous disabled message (or use the default) by supplying a blank message.