Auto login - rough working
The code here demonstrates the basic principle, but needs more work.
It handles two versions of the doi-redirect URL: the current version, and a new version that appends an extra parameter, /:anonLogin, to the end. In the second case, if anonLogin is true, then the doi-redirect controller performs an initial login.
The "sneaky" part is to realise that we should no longer perform an authentication test as part of the state transition, as this would always trigger a stateChangeError and force redirection to the login page.

It would be too fiddly to get the Anon authenticator to work in LILS, so the code here uses the simple authenticator instead. I need to decide whether to hard-wire the anonymous authenticator here, or to add yet more configuration.

It may turn out to be better to use two different URL (prefixes) rather than alternative sets of parameters.