Make DOI auth details configurable, defaulting to Anon
Allow DOI auth details to be added to topcat.json, but default to "anon" in the code. Note that specifying (non-empty) auth credentials in topcat.json is a security risk - a crafty user will be able to inspect the configuration in the browser console. So this should NOT BE USED IN PRODUCTION INSTANCES.
I have done this as the development LILS facility does not support anonymous authentication, so it is useful as a developer to be able to override the default with something else.