Attempt to rewrite CacheRepositoryTest
Following https://stackoverflow.com/a/20635285, tried to test CacheRepository using an embedded container. However, it does not appear to work. It may be that CacheRepository has to be declared explicitly using @Entity, and there may be other configuration changes required (e.g. to persistence.xml). All in all, this is turning out not to be the quick fix that I had hoped it would be, so I am using a branch to "archive" the attempt. It is unlikely that I will return to it!