Attempt to use Arquillian for CacheRepositoryTest
Arquillian is already added as a project dependency, but does not appear to be used. I thought I would try to use it to test CacheRepository with an embedded container. However, I cannot get it to work.
Initially, this version of the class did run (during the maven build's test phase), but failed with:

SEVERE: Error while expanding archive file
java.io.FileNotFoundException: C:\Users\br54\AppData\Local\Temp\gfembed7200006165525236825tmp\applications\classes\license\LICENSE (The system cannot find the path specified)
...
	at javax.ejb.embeddable.EJBContainer.createEJBContainer(EJBContainer.java:102)
	at org.icatproject.topcat.CacheRepositoryTest.setUp(CacheRepositoryTest.java:23)
	at sun.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
...

which means nothing to me. I tried replacing @Inject with @EJB, but this resulted in a compilation failure ("cannot find symbol") at line 21. Reverting the code still results in the same error!

So I am "archiving" this change to a branch. I may return to it someday, but probably not!