Initial UserResourceTest class added
UserResourceTest is not a unit test. Like IcatClientTest, it requires an existing ICAT instance; but also uses Arquillian to provide an embedded container.
I am somewhat unsatisfied with this: most notably, it requires a copy of topcat.properties to be present at the project root - all attempts to load it via resources have failed. (It may be that Properties.java should be rewritten to load topcat.properties as a resource rather than a file.)

The class contains just one simple test, that a getSize request for an Investigation returns a positive value.