Add tests for UserResource Cart methods
Added a series of embedded-container tests of the UserResource Cart methods that empty the cart, and add and delete a cart item, each time testing the size of the resulting cart. All tests use the UserResource API; it could be arged that we should be inspecting and manipulating the Cart DB directly around separate tests of each method.

One oddity revealed through these tests is that the getCart() method is possibly inconsistent: it appears that an "empty cart" response must be parsed as JSON and cannot be cast directly to a Cart, but for the non-empty case, the exact opposite is true! And of course it's not possible to tell which case applies without attempting to obtain the Cart first... The test code contains print statements that demonstrate this. This inconsistency doesn't seem to cause any problems for the Javascript that handles the response values, so I'm unwilling to (try to) change it.