Add submitCart tests
testSubmitCart() is effectively an integration test that runs through a mini-scenario. It populates the cart, then submits it, then checks that a new download has been added with expected properties. Next it changes the download status, then flags the download as deleted.
Note that this test does not clean up after itself - there will be one more download in the DB. (There is no API to truly remove a download.) So at the moment it cannot assume that there are no pre-existing downloads.