Change download API tests to use new dummy download
The previous version assumed that there was at least one existing download to use in the test; this turns out not to be the case in Travis. Using a new dummy download is better, though it required a new removeDownload() method in DownloadRepository to allow us to clean up afterwards.

In the test of setDownloadTypeStatus, reset the status (if any) to the initial value at the end of the test; but note that if there was no status entry initially, we should set the disabled status to false now!
(Perhaps it would be better to test using a dummy test downloadType...)