Add topcat_admin_LILS to distribution, build and setup
The original topcat_admin script cannot be used with the installed server-only code, as it tries to read topcat.json. topcat_admin_LILS is a variant that hard-wires local properties that would have been read from topcat.json, so can be (only) used with the local server.
The original script can still be used to connect to other Topcat instances, if desired, so has not been removed.