Refactor StatusCheck for testing
Pulled out the body of the scheduled poll() method into a separate method that can be called in testing; the new method takes an optional IdsClient, which could be replaced by a mock in tests.

It may prove necessary to refactor further, and remove the scheduled method to a separate class; or find another way to disable it when testing with an embedded container.