Add StatusCheckTest (not yet working)
Made StatusCheck.updateStatuses protected (introspection did not seem to work properly - commented-out in this version of StatusCheckTest.)
However, the method invocation fails with a NullPointerException, apparently on the line itself; I have yet to discover why.
I'm committing this version as I'm about to try some radical changes, and may want to come back to this!