Giving up: mail Session prevents embedded testing
As committed here, the test fails with "A system exception occurred during an invocation on EJB StatusCheck". Temporarily removing the @Schedule in StatusCheck reveals that the real problem is with the mail Session and @Resource(name = "mail/topcat") which results in javax.naming.NameNotFoundException: mail.

It seems that we would have to mock the mail Session class, and the only promising approach to this I've found requires adding an interface and local classes to implement it using the real Session and a mock. At the moment this feels like too great a change to the code "just" for testing, so I am minded to abandon this attempt.