Allow test configurations to disable scheduled Download checks
If the property test.disableDownloadStatusChecks is set to true in topcat.properties, then scheduled Download checks will not take place; so there is no risk of them interfering with tests that perform manual checks.
Obviously, this should only be used in test environments, and the default value is false. It is set in the topcat.properties at the top-level of the project, which is used by the embedded-container tests.