Added test for non-http[s] two-tier downloads
For a non-http[s] download (e.g. globus), StatusCheck will never call IdsClient.isPrepared(); it is assumed that another service (e.g. pollcat) will update the download status once it is complete. This test models such a process. Pollcat uses the Topcat REST API to update the download's status, and this uses the DownloadRepository (unlike Topcat's submitCart(), which works directly with the EntityManager); so we do likewise here.

This completes the main ways in which StatusCheck is used.  However, the mail-handling has not been tested; we would probably have to add a mock email session handler to do this.