Allow markup in download type descriptions
Replaced original angularjs expression with ng-bind-html.
Tried it with this (under globus download type in topcat.json):

"description": "Use Globus to download your data, but with lots of words just to see what happens. And even more! Use Globus to download your data, but with lots of words just to see what happens. And even more! Use Globus to download your data, but with lots of words just to see what happens. And even more! <br/> And what about a <a href=\"http://google.co.uk\" target=\"_blank\">link</a>?"

and both the break and the link work.
Note the use of the target attribute in the link - without it, the link opens in the same window, and the open dialog is lost when we go back.