Fix cart download disabling behaviour
Ensure that the Download Cart button is disabled if count/size limits are exceeded (and if limits are to be enforced).
Treat the count/size maximum values as valid. (The cart template HTML already does this, and the disagreement was confusing.)

Fixes #459.

This commit includes debug logging; I will remove that in the next commit.