Report cart submission failure to user
... and don't say a new download was added when it wasn't!

submitCart failure is already logged to the console; added a 'throw' at the same point. (As the comment says, this is a bit extreme, but I think it's simpler than most other mechanisms for getting a message to the user at this point!)

Use a someDownloadWasCreated flag so that the 'cart:submit' broadcast only happens when at least one submitCart() call succeeded. The broadcast is used to trigger the 'new download was created' tooltip.