StatusCheck forgives all but TopcatExceptions
Both performCheck and prepareDownload now only expire the download when a TopcatException is encountered. For all other exceptions, the lastChecks timestamp for the download is updated, so that StatusCheck does not immediately re-try.
Note that this re-try delay (configured by poll.interval.wait) is now also applied to the initial prepareData call. I am assuming this is preferable to having Topcat try to send prepareData requests to the IDS once per second until it either succeeds or throws a TopcatException.

Fixes #462.