Add test that delays are observed even when exceptions are thrown
The new test is very similar to the existing testDelays() method, except that the mock client is set to throw an IOException when isPrepared is called.
The most important test here is that after the exception has been thrown, the mock IdsClient should not be called again until the interval has passed.